:art: moves TRACKING_EVENTS constant to common/constants
