+++ src/components/ShareSaveQuery/ShareQuery.js
@@ -13,7 +13,8 @@ import Tooltip from 'uikit/Tooltip';
 import { shortUrlResolveRoot } from 'common/injectGlobals';
 import shortenApi from './shortenApi';
 import { Trans } from 'react-i18next';
-import { trackUserInteraction, TRACKING_EVENTS } from '../../services/analyticsTracking';
+import { TRACKING_EVENTS } from 'common/constants';
+import { trackUserInteraction } from '../../services/analyticsTracking';
 import { ButtonContainer, CustomLightButton } from './ui';
 import { WhiteButton } from 'uikit/Button';
 import styled from 'react-emotion';
